Extract backreference treatment to a separate module
